+++ relay-server/src/actors/processor.rs
@@ -1065,8 +1065,16 @@ impl EnvelopeProcessorService {
 
     /// Remove profiles from the envelope if they can not be parsed
     fn filter_profiles(&self, state: &mut ProcessEnvelopeState) {
+        let transaction_count: usize = state
+            .managed_envelope
+            .envelope()
+            .items()
+            .filter(|item| item.ty() == &ItemType::Transaction)
+            .count();
         let mut found_profile = false;
         state.managed_envelope.retain_items(|item| match item.ty() {
+            // Drop profile without a transaction in the same envelope.
+            ItemType::Profile if transaction_count == 0 => ItemAction::DropSilently,
             ItemType::Profile => {
                 if !found_profile {
                     match relay_profiling::parse_metadata(&item.payload()) {